Fix #7 - use SpringApplicationRunListener to load external configuration instead of a trait
