+++ src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.SpringApplicationRunListener=grails.plugin.externalconfig.ExternalConfigRunListener
